+++ src/index.ts
@@ -12,14 +12,15 @@ import onResponse from "./hooks/onResponse.ts"
 import preHandler from "./hooks/preHandler.ts"
 import launcher from "./launcher.ts"
 
+export { createKeycloakVerifier } from "./auth/keycloak.ts"
 export type {
-    ApiAuthConfig,
-    ApiAuthValidator,
     DefaultPluginsOptions,
     DefaultRoutesOptions,
     FSTPlugin,
+    KeycloakAuthConfig,
     LauncherLocals,
     LauncherOptions,
+    TokenVerifier,
 } from "./types.ts"
 
 export {